project/backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
psycopg2-binary==2.9.6